requirements.txt
asgiref==3.5.2
Django==4.0.6
django-extensions==3.2.0
django-filter==22.1
sqlparse==0.4.2